Fix assigning hint characters to windows
Due to the use of the modulo operator, hinting characters would start at
the second character instead of the first one. This is the result of Lua
using 1-indexed arrays, but the modulo operator "starting" at 0. For
example:

    1 % 4 => 1
    2 % 4 => 2
    3 % 4 => 3
    4 % 4 => 0

This commit fixes window_keys() by taking a more manual approach to
assigning hint characters, ensuring the first window uses the first hint
character instead of the second one.