Fix hinting with multiple characters
The previous implementation would always jump to the first window that
matches, even when more matches were present. In this new setup, if
there are windows that have a common prefix but contain extra hints, new
floating windows are opened that show only the overlapping hints. To
illustrate, consider the following hint options:

1. a
2. r
3. ra

When you press "r", nvim-window removes the float with the "a" hint. If
you then press "r" again it picks window 2. If you instead pressed "a",
it would have picked window 3.