Revert "Detect non-focusable container windows"
This change makes it more difficult to switch between windows when
there's a floating window present, something I didn't really notice
until recently. For example, consider two splits in which one has a
floating window (e.g. an LSP popup of some sort):

            A              B
    +---------------+---------------+
    |               |               |
    |               | +-------+     |
    |               | | float |     |
    |               | +-------+     |
    |               |               |
    +---------------+---------------+

If split window B is focused, nvim-window would allow you to jump to
either A or the floating window. However, if window A is active you
wouldn't be able to jump back to B but only the float, because window B
would be removed from the list.

In addition, for small floating windows (i.e. most of them) there's no
sensible way to display the hint icon: drawing a floating window on top
will likely result in it being far too big (as they have a fixed
height), and there's no statusline/winbar you can render to instead. In
theory we could fall back to using a floating window with the right size
regardless of what the `render` setting is set to, but this results in
additional complexity and a messy UI, both of which I'd like to avoid.

Since jumping in and out of floating windows can easily be done using
the standard window navigation keys, I'm hereby reverting the commit
such that window jumping works properly again, regardless of the
presence of floating windows.

This reverts commit 93af78311e53919a0b13d1bf6d857880bb0b975d.