Close floating windows before showing hints
Floating windows tend to be small (i.e. hover windows) and have a
tendency to "stick" when moving out of the window that they are relative
to using nvim-window. In addition, you can't move into these windows
using the usual window movements anyway.

To handle this, we now close floating windows before presenting the
window hints.